#52 Moved counting logic to voting service and refactored
